refactor: eliminate code duplication in test utilities
- Created centralized test utilities in src/test-utils/
- Consolidate duplicated mock creation functions across test files
- Extracted common test helper patterns to eliminate Codacy duplication issues
- Updated notification preferences to match interface requirements
- Maintained test coverage while improving code organization

CLOSES: #349
