+++ codacy-local-latest.json
@@ -1,0 +1,8 @@
+
+/home/doug/parallel-dev/dnd-tracker-next-js/src/lib/hooks/useCollaborators.ts
+  0:0  warning  File ignored because of a matching ignore pattern. Use "--no-ignore" to disable file ignore settings or use "--no-warn-ignored" to suppress this warning
+
+/home/doug/parallel-dev/dnd-tracker-next-js/src/lib/hooks/useEditableContent.ts
+  0:0  warning  File ignored because of a matching ignore pattern. Use "--no-ignore" to disable file ignore settings or use "--no-warn-ignored" to suppress this warning
+
+✖ 2 problems (0 errors, 2 warnings)
