+++ src/app/api/users/[id]/profile/__tests__/test-helpers.ts
@@ -1,13 +1,6 @@
 // Re-export shared API test helpers to maintain backward compatibility
 export {
   SHARED_API_TEST_CONSTANTS,
-  createMockSession,
-  createMockParams,
-  createMockRequest,
-  createMockUser,
-  createRequestBody,
-  expectSuccessResponse,
-  expectErrorResponse,
   expectAuthenticationError,
   expectAuthorizationError,
   expectValidationError,
@@ -18,5 +11,14 @@ export {
   executeTestRequest,
 } from '@/lib/test-utils/shared-api-test-helpers';
 
-// Legacy constant for backward compatibility
-export const TEST_USER_ID = '507f1f77bcf86cd799439011';
+// Re-export test utilities from centralized location to eliminate code duplication
+export {
+  TEST_USER_ID,
+  createMockSession,
+  createMockParams,
+  createMockRequest,
+  createMockUser,
+  createRequestBody,
+  expectSuccessResponse,
+  expectErrorResponse,
+} from '@/test-utils/test-helpers';
